rename configs2.py to configs.py
