+++ src/db/mysql.py
@@ -1,6 +1,6 @@
 import aiomysql
 
-from configs2 import db_conf
+from configs import db_conf
 
 from .base_db import BasePlaceHolderGen
