+++ src/main.py
@@ -12,7 +12,7 @@ from blueprint_app import blueprint_app
 from blueprint_auth import blueprint_auth
 from blueprint_moderation import blueprint_moderation
 from blueprint_search import blueprint_search
-from configs2 import QuartConfig, app_conf
+from configs import QuartConfig, app_conf
 from db import prime_db_pool, close_db_pool
 from moderation.api import init_moderation_db
 # from limiter import limiter
